+++ package.json
@@ -70,6 +70,7 @@
     "dotenv": "^9.0.2",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.10.18",
     "node-cron": "^3.0.0",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
